+++ server/route/index.ts
@@ -1,5 +1,6 @@
 import blogRoute from './blog'
 import authRoute from './auth'
+import { NextResponse } from 'next/server'
 
 const route = { ...blogRoute, ...authRoute }
 
@@ -10,8 +11,13 @@ type ServerSideAPIInterface = typeof serverSlideAPI
 
 type RouteKey = keyof ServerSideAPIInterface
 
-type RouteString<T extends RouteKey> = `/api/${T}`
+type RouteString<T extends RouteKey> = `${T}`
+
+interface RequestHooks {
+  getResponseInit?: () => ResponseInit
+  afterResponseHandler?: (response: NextResponse) => void | Promise<void>
+}
 
 // export default route
-export { serverSlideAPI }
+export { serverSlideAPI, type RequestHooks }
 export type { ServerSideAPIInterface, RouteKey, RouteString }