+++ utils/container.ts
@@ -1,0 +1,3 @@
+export function container(func: () => any | Promise<any>) {
+  func()
+}